Modularizacion Estadisticas, reformateo y refactorización de los resultados
